ToastsProvider useToasts refactoring
